+++ packages/router/test/bindings/contractReader/index.spec.ts
@@ -6,10 +6,9 @@ import {
   delay,
   getRandomBytes32,
   RequestContextWithTransactionId,
-  VariantTransactionData,
 } from "@connext/nxtp-utils";
 import { reset, restore, SinonStub, stub } from "sinon";
-import { providers } from "ethers/lib/ethers";
+import { providers } from "ethers";
 
 import {
   ActiveTransaction,