+++ packages/txservice/src/provider.ts
@@ -17,13 +17,14 @@ import {
   parseError,
   RpcError,
   ServerError,
-  TimeoutError,
+  OperationTimeout,
   TransactionReadError,
   TransactionReverted,
   ProviderCache,
   ReadTransaction,
   SyncProvider,
   OnchainTransaction,
+  StallTimeout,
 } from "./shared";
 
 const { FallbackProvider } = providers;
@@ -251,7 +252,7 @@ export class ChainRpcProvider {
         await this.wait(remainingConfirmations);
       }
     }
-    throw new TimeoutError({
+    throw new OperationTimeout({
       targetConfirmations: confirmations,
       remainingConfirmations,
       reverted,
@@ -609,22 +610,22 @@ export class ChainRpcProvider {
     if (needsSigner) {
       this.checkSigner();
     }
-    const errors: any[] = [];
-    const syncedProviders = this.shuffleSyncedProviders();
-    for (const provider of syncedProviders) {
+    const errors: NxtpError[] = [];
+    const shuffledProviders = this.shuffleSyncedProviders();
+    for (const provider of shuffledProviders) {
       try {
         return await method(provider);
       } catch (e) {
+        // TODO: With the addition of SyncProvider, this parse call may be entirely redundant. Won't add any compute,
+        // however, as it will return instantly if the error is already a NxtpError.
         const error = parseError(e);
-        // If the error thrown is a timeout or non-RPC or non-Server error, we want to go ahead and throw it.
-        // e.g. a TransactionReverted, TransactionReplaced, etc.
-        if (
-          error.type !== ServerError.type &&
-          (error.type !== RpcError.type || (error as RpcError).reason === RpcError.reasons.Timeout)
-        ) {
-          throw error;
-        } else {
+        if (error.type === ServerError.type || error.type === RpcError.type || error.type === StallTimeout.type) {
+          // If the method threw a StallTimeout, RpcError, or ServerError, that indicates a problem with the provider and not
+          // the call - so we'll retry the call with a different provider (if available).
           errors.push(error);
+        } else {
+          // e.g. a TransactionReverted, TransactionReplaced, etc.
+          throw error;
         }
       }
     }
@@ -654,7 +655,11 @@ export class ChainRpcProvider {
 
     // First, sync all providers simultaneously.
     await Promise.all(
-      this.providers.map((p) => new Promise((resolve, reject) => p.sync().then(resolve).catch(reject))),
+      this.providers.map(async (p) => {
+        try {
+          await p.sync();
+        } catch (e) {}
+      }),
     );
 
     // Find the provider with the highest block number and use that as source of truth.
@@ -678,19 +683,24 @@ export class ChainRpcProvider {
 
     // We want to pick the lead provider here at random from the list of 0-lag providers to ensure that we distribute
     // our block listener RPC calls as evenly as possible across all providers.
-    const leadProviders = this.shuffleSyncedProviders(true);
-    this.leadProvider = leadProviders[(Math.random() * leadProviders.length) | 0];
+    const leadProviders = this.shuffleSyncedProviders();
+    this.leadProvider = leadProviders[0];
 
     this.logger.debug("Synced provider(s).", requestContext, methodContext, {
       highestBlockNumber,
       leadProvider: this.leadProvider.name,
-      providers: this.providers
-        .filter((p) => p.synced)
-        .map((p) => ({
-          url: p.name,
-          blockNumber: p.syncedBlockNumber,
-          lag: p.lag,
-        })),
+      providers: this.providers.map((p) => ({
+        url: p.name,
+        blockNumber: p.syncedBlockNumber,
+        lag: p.lag,
+        synced: p.synced,
+        metrics: {
+          reliability: p.reliability,
+          latency: p.latency,
+          cps: p.cps,
+          priority: p.priority,
+        },
+      })),
     });
   }
 
@@ -711,21 +721,26 @@ export class ChainRpcProvider {
    * @returns all in-sync providers in order of synchronicity with chain, with the lead provider
    * in the first position and the rest shuffled by tier (lag).
    */
-  private shuffleSyncedProviders(zeroLagOnly = false): SyncProvider[] {
-    const syncedProviders = this.providers.filter((p) => (zeroLagOnly ? p.lag === 0 : p.synced));
+  private shuffleSyncedProviders(): SyncProvider[] {
+    // TODO: Should priority be a getter, and calculated internally?
     // Tiered shuffling: providers that have the same lag value (e.g. 0) will be shuffled so as to distribute RPC calls
     // as evenly as possible across all providers; at high load, this can translate to higher efficiency (each time we
     // execute an RPC call, we'll be hitting different providers).
     // Shuffle isn't applied to lead provider - instead, we just guarantee that it's in the first position.
-    syncedProviders.forEach((p) => {
+    this.providers.forEach((p) => {
       p.priority =
         p.lag -
         (this.leadProvider && p.name === this.leadProvider.name ? 1 : Math.random()) -
         p.cps / this.config.maxProviderCPS -
-        p.reliability * 2 +
-        p.avgExecTime;
+        // Reliability factor reflects how often RPC errors are encountered, as well as timeouts.
+        p.reliability * 10 +
+        p.latency;
     });
-    return syncedProviders.sort((a, b) => a.priority - b.priority);
+    // Always start with the in-sync providers and then concat the out of sync subgraphs.
+    return this.providers
+      .filter((p) => p.synced)
+      .sort((a, b) => a.priority - b.priority)
+      .concat(this.providers.filter((p) => !p.synced).sort((a, b) => a.priority - b.priority));
   }
 
   private async setBlockPeriod(): Promise<void> {