+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Next Release
 
+# 0.0.103
+
+- [txservice] Fix handling of RPC provider server errors
+
+# 0.0.102
+
+- [txservice] Fixed issue with RPC stall timeout handling
 - [router] Undefined handling in `bindPrices`
+- [router/sdk] Get active transactions specific to users and routers
 
 # 0.0.101